Respite and Run/NoRun Update
Added the function to allow for the app to check when the assignment
should run, and when it should not run.

Also added the Respite check to ensure that the user is notified of the
assignment ahead of it's due date by the appropriate number of hours and
minutes. This is checked against the InternalClock.