Added PreStatus, Fixed textBrowser, and more
Added the PreStatus parsing to ensure that PreStatuses are being
respected so that functions are only available, or are only ran when in
specific Statuses. This fixed issues where all Reports available in the
.ini script were being displayed, even when they should not be.

Replaced the textBrowser with a ScrollArea to fix an issue where the
textBrowser would scroll back to the top when the textBrowser refreshed
to show differences in the displayed text.

Added the DataInspectorDialog and added it to the Test Menu as 'View
Script Data' to allow users to view the parsed script data and verify
it's correct.

Fixed the question parsing to ensure that questions are being displayed
properly and are correctly setting variables, and correctly calling
procedures.

Updated runProcedure to ensure that it's properly running sub-commands
in the order that they are presented in the .ini files. This fixed
Mathing issues for procedures that contain math such as Add, Subtract,
Multiply, Divide, Greater Than (>), and Less Than (<).

Refactored all aspects of the cyberdom.cpp file to pull data from the
ScriptData.h structures instead of pulling the data directly from the
raw parsed .ini data, allowing for the data to be pulled and utilized in
a structured format.

Updated the new ScrollArea to display TopText= values from the [General]
section of the .ini and Text= values of Jobs and Punishments. TopText=
values will always appear at the top of the ScrollArea and Text= values
will always appear at the middle of the ScrollArea. BottomText= values
will appear at the bottom of the ScrollArea, beneath any Text= values.

Updated the ScriptData.h file to include PreStatus in the
ConfessionDefinition so that PreStatus can be utilized to make sure that
Confessions are visible in the Communication menu only when in the
correct status.