Variables and Add Clothing
Resolved issues in the addclothing.cpp that was causing double messages of "Please add a name for the item to appear" in instances where the user accidentally left the name of a new clothing item empty.

Updated the cyberdom.cpp and assignments.cpp files to fix an issue that was causing the {!zzRunTime} and {!zzMinTime} variables to not be replaced with the correct values.