Reports and Logs
Updated the reports path to save the HTML reports in the Documents/CyberDom/Reports folder. This change is for all Operating Systems.

Updated the Logs to be saved in the Documents/CyberDom/Logs folder. This applies for all Operating Systems.

Updated the viewReportFile function to use "xdg-open" to use the default browser on Linux systems as Linux doesn't work well with the default Qt method of pulling the default browser.