Icons, Job Text, Version Number, Logs
Fixed an issue with icons not showing on some Linux Desktops.

Resolved an issue with Job Text being duplicated in the Assignment Notes and CyberDom ScrollArea.

Updated the logic for the Version Number to better automate the change to the Version Number.

Updated the Logs to provide more detailed information about the device that the application is being ran on.